cambio en las migraciones para poner las foreign keys como INT y no BIGINT

+++ database/migrations/2024_01_07_172218_create_horarios_clases_table.php
@@ -11,17 +11,17 @@ return new class extends Migration {
     public function up(): void {
         Schema::create('horarios_clases', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('horario_id');
+            $table->unsignedInteger('horario_id');
             $table->foreign('horario_id')->references('id')->on('horarios');
-            $table->unsignedBigInteger('dia_semana_id');
+            $table->unsignedInteger('dia_semana_id');
             $table->foreign('dia_semana_id')->references('id')->on('dias_semana_tabla_maestra');
-            $table->unsignedBigInteger('franja_horaria_id');
+            $table->unsignedInteger('franja_horaria_id');
             $table->foreign('franja_horaria_id')->references('id')->on('franjas_horarias_tabla_maestra');
-            $table->unsignedBigInteger('clase_id');
+            $table->unsignedInteger('clase_id');
             $table->foreign('clase_id')->references('id')->on('clases');
-            $table->unsignedBigInteger('monitor_id');
+            $table->unsignedInteger('monitor_id');
             $table->foreign('monitor_id')->references('id')->on('personal');
-            $table->unsignedBigInteger('sala_id');
+            $table->unsignedInteger('sala_id');
             $table->foreign('sala_id')->references('id')->on('salas');
             $table->timestamps();
         });
